rename DrawRectangle method into Draw
